feat: 뷰와 로직 분리
